Moved AtomicInt and encapsulated CurrentTime methods into an object
