+++ build.gradle.kts
@@ -23,7 +23,8 @@ version = "1.0-SNAPSHOT"
 plugins {
     kotlin("multiplatform") version "1.7.0"
 
-    id("com.diffplug.spotless") version "6.9.0"
+    id("com.diffplug.spotless") version "6.10.0"
+    id("io.gitlab.arturbosch.detekt").version("1.21.0")
     id("org.jetbrains.dokka") version "1.7.10"
 }
