+++ src/jvmTest/kotlin/org/kbson/BsonValueApiTest.kt
@@ -16,11 +16,11 @@
 package org.kbson
 
 import java.lang.reflect.Constructor
+import java.util.Collections
+import java.util.UUID
 import kotlin.jvm.internal.DefaultConstructorMarker
 import kotlin.test.assertEquals
 import org.junit.jupiter.api.Test
-import java.util.Collections
-import java.util.UUID
 
 class BsonValueApiTest {
